[Refactor] DbContext의 namespace가 올바르지 않던 문제 해결
