+++ server/Program.cs
@@ -1,9 +1,9 @@
-﻿using server.Data;
+﻿using server;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-builder.Services.AddScoped<ApplicationDbContext>();
+builder.Services.AddDbContext<ApplicationDbContext>();
 
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
